GH-139: Improve preflight config validation
Depends on https://github.com/confluentinc/kafka-connect-bigquery/pull/140

Addresses the remainder of https://github.com/confluentinc/kafka-connect-bigquery/issues/139.

All multi-property configuration validation is moved from the constructor of the BigQuerySinkConfig class
to a new `BigQuerySinkConfig::validate` method, which is invoked from `BigQuerySinkConnector::validate` and
leverages the Connect API for defining custom error messages on a per-property basis instead of throwing
exceptions on invalid configurations, which only displays one message at a time and doesn't work well with
programmatic UIs.

Unit tests are added for all new validation logic.

The logic for GCP client construction is also updated to conform to a common interface, which should
improve readability and maintainability, and makes it easier to validate credentials for either BigQuery or GCS.

Finally, the `SinkConfigConnectException` class is removed as it's not really necessary and doesn't bring anything
worthwhile to the code base.